utils - folderInfo (independent functions)
